+++ spinta/dimensions/comments/components.py
@@ -19,3 +19,5 @@ class Comment(Component):
     created: str
     comment: str
     given: CommentGiven
+    prepare: str | None = None
+    level: int | None = None